+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Not yet released.
 
 * [CS-542][] Add `"container-user"` to Docker Server settings. This allows you to specify the user within the container as whom the command should be run. If you specify nothing, the process within the container is run as root or whatever user settings are defined on the image (which is unchanged from the current behavior).
 * [CS-543][] Add support for setting container user as `"containerUser=username"` in prefs-init.ini config file. This value is only used when no other server settings exist in the database, such as initial deployment. 
+* [CS-547][] Setup and wrapup container environment variables are now populated with resolved values from parent command. 
 
 ### Bugfixes
 
@@ -92,6 +93,7 @@ Not yet released.
 [CS-531]: https://issues.xnat.org/browse/CS-531
 [CS-535]: https://issues.xnat.org/browse/CS-535
 [CS-537]: https://issues.xnat.org/browse/CS-537
+[CS-547]: https://issues.xnat.org/browse/CS-547
 
 ## 1.5.1